More subtle update notif, code cleanup
