+++ packages/toolpad-app/src/toolpad/ToolpadShell/index.tsx
@@ -1,0 +1,31 @@
+import * as React from 'react';
+import { styled } from '@mui/material';
+import Header from './Header';
+
+export interface ToolpadShellProps {
+  actions?: React.ReactNode;
+  status?: React.ReactNode;
+  children?: React.ReactNode;
+}
+
+const ToolpadShellRoot = styled('div')({
+  width: '100vw',
+  minHeight: '100vh',
+  display: 'flex',
+  flexDirection: 'column',
+});
+
+const ViewPort = styled('div')({
+  flex: 1,
+  overflow: 'auto',
+  position: 'relative',
+});
+
+export default function ToolpadShell({ children, ...props }: ToolpadShellProps) {
+  return (
+    <ToolpadShellRoot>
+      <Header {...props} />
+      <ViewPort>{children}</ViewPort>
+    </ToolpadShellRoot>
+  );
+}
